requirements.txt
pandas
google-generativeai
python-dotenv
fpdf2
streamlit